+++ index.html
@@ -11,7 +11,7 @@
 
 <body>
   <main>
-    <script src="js/main.js"></script>
+    <script src="js/main.js" type="module"></script>
 
     <!-- Контейнер для изображений от других пользователей -->
     <section class="pictures  container">
